+++ tiledb/sm/cpp_api/array_schema.h
@@ -361,6 +361,84 @@ class ArraySchema : public Schema {
   }
 
   /**
+   * Returns a copy of the FilterList of the coordinates. To change the
+   * coordinate compressor, use `set_coords_filter_list()`.
+   *
+   * @return Copy of the coordinates FilterList.
+   */
+  FilterList coords_filter_list() const {
+    auto& ctx = ctx_.get();
+    tiledb_filter_list_t* filter_list;
+    ctx.handle_error(tiledb_array_schema_get_coords_filter_list(
+        ctx, schema_.get(), &filter_list));
+    return FilterList(ctx, filter_list);
+  }
+
+  /**
+   * Sets the FilterList for the coordinates, which is an ordered list of
+   * filters that will be used to process and/or transform the coordinate data
+   * (such as compression).
+   *
+   * **Example:**
+   * @code{.cpp}
+   * tiledb::Context ctx;
+   * tiledb::ArraySchema schema(ctx, TILEDB_SPARSE);
+   * tiledb::FilterList filter_list(ctx);
+   * filter_list.add_filter({ctx, TILEDB_FILTER_BYTESHUFFLE})
+   *     .add_filter({ctx, TILEDB_FILTER_BZIP2});
+   * schema.set_coords_filter_list(filter_list);
+   * @endcode
+   *
+   * @param filter_list FilterList to use
+   * @return Reference to this `ArraySchema` instance.
+   */
+  ArraySchema& set_coords_filter_list(const FilterList& filter_list) {
+    auto& ctx = ctx_.get();
+    ctx.handle_error(tiledb_array_schema_set_coords_filter_list(
+        ctx, schema_.get(), filter_list));
+    return *this;
+  }
+
+  /**
+   * Returns a copy of the FilterList of the offsets. To change the
+   * offsets compressor, use `set_offsets_filter_list()`.
+   *
+   * @return Copy of the offsets FilterList.
+   */
+  FilterList offsets_filter_list() const {
+    auto& ctx = ctx_.get();
+    tiledb_filter_list_t* filter_list;
+    ctx.handle_error(tiledb_array_schema_get_offsets_filter_list(
+        ctx, schema_.get(), &filter_list));
+    return FilterList(ctx, filter_list);
+  }
+
+  /**
+   * Sets the FilterList for the offsets, which is an ordered list of
+   * filters that will be used to process and/or transform the offsets data
+   * (such as compression).
+   *
+   * **Example:**
+   * @code{.cpp}
+   * tiledb::Context ctx;
+   * tiledb::ArraySchema schema(ctx, TILEDB_SPARSE);
+   * tiledb::FilterList filter_list(ctx);
+   * filter_list.add_filter({ctx, TILEDB_FILTER_POSITIVE_DELTA})
+   *     .add_filter({ctx, TILEDB_FILTER_LZ4});
+   * schema.set_offsets_filter_list(filter_list);
+   * @endcode
+   *
+   * @param filter_list FilterList to use
+   * @return Reference to this `ArraySchema` instance.
+   */
+  ArraySchema& set_offsets_filter_list(const FilterList& filter_list) {
+    auto& ctx = ctx_.get();
+    ctx.handle_error(tiledb_array_schema_set_offsets_filter_list(
+        ctx, schema_.get(), filter_list));
+    return *this;
+  }
+
+  /**
    * Returns a copy of the schema's array Domain. To change the domain,
    * use `set_domain()`.
    *